Use fixtures for helpers and move drift tests
